Sweep out the visibility refactor's leftover scaffolding
Phase 4 (final) of the scene visibility store plan (#579). With the store
as the only published visibility truth, remove everything the flip
obsoleted and unify the seams it exposed:

- Retire the legacy render() draw path: the cwRHIObject pure virtual,
  all six subclass bodies, and the !gathered fallback arm in gatherScene
  (dead-but-armed; its surviving bodies drew with hardcoded slot-0
  offsets and would have used the wrong camera on an offscreen job).
  Drawable::Type::Custom stays - billboards' 2D renderer gathers it.
- Stamp cwRhiFrameRenderer* on cwRHIObject at registerRenderObject,
  beside the id stamp; delete the five subclass m_frame members, eleven
  lazy backfill blocks, and their silent null no-op guards.
- Delete the dead write-only cwRhiTexturedItems::m_renderItems QPointer.
- Split the textured-items command payload from the authoring entry:
  PendingCommand now carries ItemPayload (geometry/texture/material/
  uniformBlock/modelMatrix); visible/storeGeometry/storeTexture are
  authoring state only - no visibility bit travels on any command.
- Remove the Key::id -> store-subId truncation by widening the store to
  uint64 subIds (narrowing Key::id would break cwRayHit::objectId()'s
  qulonglong API); delete the storeSubId() helper.
- Consolidate the hand-assembled position-only test geometry into
  testlib/TestGeometryBuilders.h (cwTestGeometry::points/lines/
  triangles) across ~17 test files; fixtures that assert on
  geometry.set() or test geometry construction itself keep local builds.
- Authoring-copy audit: only the three planned keeps survive
  (cwRenderObject::m_visible, m_frontState[id].visible, the line-plot
  mask) - nothing redundant remained to delete.
- Docs: stamp the two-snapshot-read-paths note at
  cwRHILinePlot::updateVisibilityBuffer; mark intersecter-plan
  follow-up 3 landed via the Phase 2 re-key and re-scope follow-up 5 to
  its registration-ordering half; mark the scene visibility plan
  COMPLETE with a Phase 4 landing note.

Tests: full C++ suite 312,750 assertions in 1,501 cases, all passed;
full QML suite 685 passed / 0 failed / 26 skipped (offscreen); targeted
visibility/pick/intersecter tags green. cavewhere-test, CaveWhere, and
cavewhere-qml-test all build clean.

+++ cavewherelib/src/cwRenderLinePlot.h
@@ -101,7 +101,7 @@ public:
     // The plot is one geometry blob, so it occupies a single sub-slot in both
     // the intersecter and the visibility store. Shared with cwRHILinePlot's
     // snapshot reads.
-    static constexpr uint32_t kSubId = 0;
+    static constexpr uint64_t kSubId = 0;
 };
 
 inline float cwRenderLinePlot::maxZValue() const
